+++ api/photo/get_photo.js
@@ -1,0 +1,1 @@
+//TODO: 从数据库中获得对应图片id并访问接口获得无水印原图
